requirements.txt

langchain
langchain-core
langchain-community
ollama
streamlit
python-dotenv
